// apps/odonto-front/src/generated/ts/PatientResponseDto.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/


export type PatientResponseDto = {
    /**
     * @type number
    */
    id: number;
    /**
     * @type string
    */
    name: string;
    /**
     * @type string | undefined, date
    */
    birthDate?: string;
    /**
     * @type string | undefined
    */
    email?: string;
    /**
     * @type string | undefined
    */
    phone?: string;
    /**
     * @type string | undefined
    */
    address?: string;
    /**
     * @type string | undefined
    */
    document?: string;
    /**
     * @type number
    */
    clinicId: number;
    /**
     * @type string, date-time
    */
    createdAt: string;
    /**
     * @type string, date-time
    */
    updatedAt: string;
    /**
     * @type string | undefined, date-time
    */
    lastProcedureDate?: string;
    /**
     * @type string | undefined, date-time
    */
    nextAppointmentDate?: string;
};